Warning! Experimental code active on TeleOpMain
Lift code is functioning much better. We've implemented a new test clause in the LiftMotor method the lets the motor run to position when raised above the threshold for lift height. Testing halted due to a hardware malfunction.

The lift code has been activated on TeleOpMain! Gamepad 1 is safe to use with no *known bugs*. Gamepad 2 is unsafe at this time as the lift controls are active. The lift code works as intended in the downwards direction. The upwards limit has not yet been properly coded. The lift still is not able to hold its position with the exception of certain heights being held by the elastics.