Add to main to centralize qualifier code (#5)
* Corrected direction of motors

Added UP, DOWN, FORWARD, and REVERSE directions within PowerPlayRobot

* Prototype for lift holding

Changes will need to be made. There are todo comments in the code that should be acted on as soon as possible.

* Pushing Helen's code and added autonomous strafing and rotation

* Lift testing code complete. Finished grabber test.

The LiftTest code is complete. We need to finish the test on the field so we know what power needs to be applied to achieve a height.

GrabberTest code needs to be implemented into TeleOpMain.

Autonomous rotation and strafing ported over from FreightFrenzy repo.

* Tested and fixed rotate for autonomous

* Push

* Completed basic grabber functionality within GrabberTest

* Added GrabberTest components into TeleOpMain. Lift disabled TeleOpMain.

* LiftTestChanges

* Warning! Experimental code active on TeleOpMain

Lift code is functioning much better. We've implemented a new test clause in the LiftMotor method the lets the motor run to position when raised above the threshold for lift height. Testing halted due to a hardware malfunction.

The lift code has been activated on TeleOpMain! Gamepad 1 is safe to use with no *known bugs*. Gamepad 2 is unsafe at this time as the lift controls are active. The lift code works as intended in the downwards direction. The upwards limit has not yet been properly coded. The lift still is not able to hold its position with the exception of certain heights being held by the elastics.

* Lowered the threshold for testing.

* Reverted lift code to depend on RUN_USING_ENCODER

Lift code no longer uses RUN_TO_POSITION. Using the latter was causing the motor to jump between modes quickly and handicapped the lift from moving properly. This has still NOT been tested. It is NOT recommended to use Gamepad2 when using TeleOp. The lift may break if used incorrectly during this testing stage.

* Switched lift limits to use potentiometer values

* Configured liftMotor to ramp its speeds